requirements.txt
flask>=3.0
python-dotenv>=1.0
openpyxl>=3.1
gunicorn>=21.0
